fix(fuzz): commit Cargo.lock pinning getrandom 0.3.4 + repair roundtrip types
Two fixes that together unbreak the `fuzz-build` CI job.

1. `fuzz/Cargo.lock` is now committed and pins the transitive
   `getrandom` dependency to `0.3.4`. The `0.4.x` series declares
   `edition = "2024"` in its manifest and `rust-version = "1.85"`,
   both unparseable by cargo 1.82 (the cargo bundled with the
   nightly-2024-08-01 toolchain that CI uses for fuzz builds). This
   mirrors the workspace-level pin documented in `AGENTS.md` § Toolchain
   pinning. `getrandom` is only pulled in via `tempfile`, which only
   the fuzz harness uses; no production code path touches it.

   CI now runs `cargo fuzz build -- --locked` so the committed lock
   is enforced rather than silently regenerated.

2. `fuzz/fuzz_targets/roundtrip.rs` was comparing `usize::len()`
   against `MAX_KEY_LEN` / `MAX_PAYLOAD_LEN`, which are `u32`. Plain
   `cargo check --workspace` does not catch this because `fuzz/` is
   `exclude`d from the workspace; the type error only surfaced once
   cargo-fuzz actually compiled the harness on the nightly job. Cast
   both constants to `usize` at the comparison site.

`fuzz/README.md` updated to match the new CI invocation and document
the lockfile pin.

+++ fuzz/fuzz_targets/roundtrip.rs
@@ -47,12 +47,12 @@ fuzz_target!(|data: &[u8]| {
     // Clamp to MAX_KEY_LEN; the rest becomes the payload.
     let body = &data[2..];
     let key_len_raw = u16::from_le_bytes([data[0], data[1]]) as usize;
-    let key_len = key_len_raw.min(body.len()).min(MAX_KEY_LEN);
+    let key_len = key_len_raw.min(body.len()).min(MAX_KEY_LEN as usize);
 
     let key = &body[..key_len];
     let payload = &body[key_len..];
 
-    if payload.len() > MAX_PAYLOAD_LEN {
+    if payload.len() > MAX_PAYLOAD_LEN as usize {
         return;
     }
